core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@4883 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/TextTable.java
@@ -135,7 +135,7 @@ public class TextTable extends org.hsqldb.Table {
 
                 Object[] data = row.getData();
 
-                nextpos = row.getPos() + row.getStorageSize();
+                nextpos = (int) row.getPos() + row.getStorageSize();
 
                 systemUpdateIdentityValue(data);
                 enforceRowConstraints(session, data);
